Renamed pages and created router
