+++ frontend/src/pages/Home.test.tsx
@@ -2,7 +2,7 @@
 /* eslint-disable global-require */
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
-import { Home } from './Home';
+import { Home } from './Login';
 
 jest.mock('react-router-dom', () => ({
   Navigate: () => <div>Navigate</div>,
