+++ frontend/src/pages/Login.tsx
@@ -1,0 +1,22 @@
+import { Navigate } from 'react-router-dom';
+import { AuthenticationDialog, Dialog } from '../components';
+import { useAuthentication } from '../hooks';
+
+export const Home = () => {
+  const { isAuthenticated, hasRole } = useAuthentication();
+
+  if (isAuthenticated) {
+    if (hasRole('admin') || hasRole('hm') || hasRole('shr') || hasRole('adm'))
+      return <Navigate replace to="Home" />;
+    return (
+      <Dialog
+        title="You have not been assigned one on the necessary roles to access this application."
+        open
+      />
+    );
+  }
+
+  return <AuthenticationDialog open title="Authentication Required" />;
+};
+
+export default Home;
