+++ frontend/src/pages/index.ts
@@ -1,3 +1,3 @@
+export * from './Login';
 export * from './Home';
-export * from './Inquiry';
 export * from './Admin';
